requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
sqlalchemy==2.0.35
pillow>=11.0.0
python-multipart==0.0.9
httpx==0.28.1

